REVSDL-2506 Implement JSON notifications
