+++ src/components/vr_cooperation/include/vr_cooperation/commands/on_default_service_chosen_notification.h
@@ -34,7 +34,7 @@
 #define SRC_COMPONENTS_VR_COOPERATION_INCLUDE_VR_COOPERATION_COMMANDS_ON_DEFAULT_SERVICE_CHOSEN_NOTIFICATION_H_
 
 #include <string>
-#include "vr_cooperation/commands/base_command_notification.h"
+#include "vr_cooperation/commands/json_notification.h"
 #include "vr_cooperation/interface/hmi.pb.h"
 
 namespace vr_cooperation {
@@ -46,7 +46,7 @@ namespace commands {
 /**
  * @brief OnDefaultServiceChosenNotification command class
  */
-class OnDefaultServiceChosenNotification : public BaseCommandNotification {
+class OnDefaultServiceChosenNotification : public JSONNotification {
  public:
   /**
    * @brief OnDefaultServiceChosenNotification class constructor
