+++ src/components/vr_cooperation/src/command/on_service_deactivated_notification.cc
@@ -47,7 +47,7 @@ CREATE_LOGGERPTR_GLOBAL(logger_, "VRCooperation")
 
 OnServiceDeactivatedNotification::OnServiceDeactivatedNotification(
     VRModule* parent, const vr_hmi_api::ServiceMessage& message)
-    : BaseCommandNotification(parent, message) {
+    : JSONNotification(parent, message) {
 }
 
 void OnServiceDeactivatedNotification::Execute() {
