--- src/components/vr_cooperation/src/hmi_command_factory.cc
@@ -1,61 +1,0 @@
-/*
- * Copyright (c) 2016, Ford Motor Company
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following
- * disclaimer in the documentation and/or other materials provided with the
- * distribution.
- *
- * Neither the name of the Ford Motor Company nor the names of its contributors
- * may be used to endorse or promote products derived from this software
- * without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-
-#include "vr_cooperation/hmi_command_factory.h"
-
-#include "utils/logger.h"
-#include "vr_cooperation/interface/hmi.pb.h"
-#include "vr_cooperation/vr_module.h"
-#include "vr_cooperation/commands/on_default_service_chosen_notification.h"
-#include "vr_cooperation/commands/on_service_deactivated_notification.h"
-
-namespace vr_cooperation {
-
-CREATE_LOGGERPTR_GLOBAL(logger_, "VRCooperation")
-
-commands::Command* HMICommandFactory::CreateCommand(
-    VRModule* parent,
-    const vr_hmi_api::ServiceMessage& message) {
-  LOG4CXX_AUTO_TRACE(logger_);
-  switch (message.rpc()) {
-    case vr_hmi_api::RPCName::ON_DEACTIVATED:
-      return new commands::OnServiceDeactivatedNotification(parent, message);
-    case vr_hmi_api::ON_DEFAULT_CHOSEN:
-      return new commands::OnDefaultServiceChosenNotification(parent, message);
-    default: {
-      return NULL;
-    }
-  }
-}
-
-}  // namespace vr_cooperation
-
